Rename Transformer interface and package
