+++ readers/api/responses.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 
 	"github.com/mainflux/mainflux"
-	"github.com/mainflux/mainflux/transformer/senml"
+	"github.com/mainflux/mainflux/transformers/senml"
 )
 
 var _ mainflux.Response = (*pageRes)(nil)
