+++ readers/cassandra/messages.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/gocql/gocql"
 	"github.com/mainflux/mainflux/readers"
-	"github.com/mainflux/mainflux/transformer/senml"
+	"github.com/mainflux/mainflux/transformers/senml"
 )
 
 var _ readers.MessageRepository = (*cassandraRepository)(nil)
