+++ readers/cassandra/messages_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/mainflux/mainflux/readers"
 	creaders "github.com/mainflux/mainflux/readers/cassandra"
-	"github.com/mainflux/mainflux/transformer/senml"
+	"github.com/mainflux/mainflux/transformers/senml"
 	cwriters "github.com/mainflux/mainflux/writers/cassandra"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
