+++ readers/mocks/messages.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 
 	"github.com/mainflux/mainflux/readers"
-	"github.com/mainflux/mainflux/transformer/senml"
+	"github.com/mainflux/mainflux/transformers/senml"
 )
 
 var _ readers.MessageRepository = (*messageRepositoryMock)(nil)
