+++ readers/mongodb/messages.go
@@ -7,7 +7,7 @@ import (
 	"context"
 
 	"github.com/mainflux/mainflux/readers"
-	"github.com/mainflux/mainflux/transformer/senml"
+	"github.com/mainflux/mainflux/transformers/senml"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
