+++ transformers/doc.go
@@ -1,6 +1,6 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-// Package transformer contains the domain concept definitions needed to
+// Package transformers contains the domain concept definitions needed to
 // support Mainflux transformer services functionality.
-package transformer
+package transformers
