+++ transformers/senml/message.go
